hanzo_message_primitives::schemas

Module wallet_mixed

Structs

Address

Represents an address in a wallet.

AddressBalanceList

Represents a list of balances for an address. For now we’ll just track ETH, USDC and KAI.

AddressList

Represents a list of addresses in a wallet.

Asset

Represents an asset onchain scoped to a particular network.

Balance

Represents the balance of an asset onchain.

CreateTransferRequest

Represents a request to create a transfer.

ModelError

Represents an error response from the Coinbase Developer Platform API.

PublicAddress

Represents an address in a wallet.

Transaction

Represents an onchain transaction.

Transfer

Represents a transfer of an asset from one address to another.

TransferList

Represents a list of transfers.

Enums

AssetType

NetworkIdentifier

Enum representing the ID of the blockchain network.

NetworkProtocolFamilyEnum

Enum representing the protocol family of the blockchain network.

TransactionStatusEnum

Enum representing the status of the transaction.

TransactionType

Enum representing the type of transaction.

TransferStatusEnum

Enum representing the status of the transfer.